@@ -2,14 +2,14 @@ apply plugin: 'com.android.application'
 
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 31
     defaultConfig {
         applicationId "com.ebe.miniaelec"
         minSdkVersion 21
         targetSdkVersion 28
-        versionCode 23 // next test version 25, production 23
-        versionName "1.0.23"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        versionCode 26 // next test version 25, production 23
+        versionName "1.0.26"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -21,29 +21,35 @@ android {
         jni.srcDirs = []
         jniLibs.srcDir "src/main/libs"
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:design:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.core:core:1.7.0'
+    //implementation 'androidx.core:core-ktx:1.7.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    implementation 'com.google.android.material:material:1.5.0'
     implementation files('libs/GLPage_V1.03.00_20181030.jar')
     implementation files('libs/NeptuneLiteApi_V3.21.00_20210120.jar')
     implementation files('libs/EBEUnifiedLibrary.jar')
     testImplementation 'junit:junit:4.12'
-    implementation 'com.squareup.retrofit2:retrofit:2.3.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.3.0'
     implementation 'com.squareup.okhttp3:okhttp:4.9.3'
     implementation 'com.github.d-max:spots-dialog:0.7@aar'
     implementation("com.squareup.okhttp3:logging-interceptor:4.9.3")
-    //RxJava RxAndroid
+    //RxJava2 RxAndroid
     //implementation 'io.reactivex.rxjava2:rxjava:2.2.2'
     //implementation 'io.reactivex.rxjava2:rxandroid:2.1.0'
     implementation 'com.tbruyelle.rxpermissions2:rxpermissions:0.9.4@aar'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
     implementation files('libs/ormlite-android-5.1.jar')
 
     //leak canary
@@ -62,5 +68,9 @@ dependencies {
     // explicitly depend on RxJava's latest version for bug fixes and new features.
     // (see https://github.com/ReactiveX/RxJava/releases for latest 3.x.x version)
     implementation 'io.reactivex.rxjava3:rxjava:3.1.3'
+
+
+    //easy permissions
+    implementation 'pub.devrel:easypermissions:3.0.0'
 
 }
