@@ -1,7 +1,7 @@
 package com.ebe.miniaelec.http;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 import android.widget.Toast;
 
