@@ -5,9 +5,9 @@ import android.content.Context;
 import android.content.res.Configuration;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.Log;
@@ -24,14 +24,12 @@ import com.ebe.miniaelec.R;
 import com.ebe.miniaelec.database.DBHelper;
 import com.ebe.miniaelec.http.ApiServices;
 import com.ebe.miniaelec.http.RequestListener;
-import com.ebe.miniaelec.model.BillData;
 import com.ebe.miniaelec.model.TransBill;
 import com.ebe.miniaelec.model.TransData;
 import com.ebe.miniaelec.print.PrintListener;
 import com.ebe.miniaelec.print.PrintReceipt;
 import com.ebe.miniaelec.utils.Utils;
 import com.google.gson.Gson;
-import com.j256.ormlite.dao.ForeignCollection;
 
 import org.json.JSONArray;
 import org.json.JSONException;
